game graphics showing
