+++ public/games/CakeGame.js
@@ -1,4 +1,4 @@
-import Phaser from 'phaser'
+// import Phaser from 'phaser'
 
 CakeGame.Game = function(game) {
     this._player = null;
